Move API modules (api, request, socket, msgpack) into their own subdirectory src/api.
